+++ app/src/main/java/org/ole/planet/myplanet/data/repository/UserRepository.kt
@@ -1,0 +1,36 @@
+package org.ole.planet.myplanet.data.repository
+
+import android.content.SharedPreferences
+import io.realm.Realm
+import org.ole.planet.myplanet.datamanager.ApiInterface
+import org.ole.planet.myplanet.datamanager.DatabaseService
+import org.ole.planet.myplanet.model.RealmUserModel
+
+interface UserRepository {
+    suspend fun getUserProfile(): String?
+    suspend fun saveUserData(data: String)
+    fun getRealm(): Realm
+    fun getCurrentUser(): RealmUserModel?
+}
+
+class UserRepositoryImpl(
+    private val databaseService: DatabaseService,
+    private val preferences: SharedPreferences,
+    private val apiInterface: ApiInterface
+) : UserRepository {
+    override suspend fun getUserProfile(): String? {
+        return preferences.getString("user_profile", null)
+    }
+
+    override suspend fun saveUserData(data: String) {
+        preferences.edit().putString("user_profile", data).apply()
+    }
+
+    override fun getRealm(): Realm {
+        return databaseService.realmInstance
+    }
+
+    override fun getCurrentUser(): RealmUserModel? {
+        return databaseService.realmInstance.where(RealmUserModel::class.java).findFirst()
+    }
+}
