+++ app/src/main/java/org/ole/planet/myplanet/di/RepositoryModule.kt
@@ -5,13 +5,15 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import io.realm.Realm
 import javax.inject.Singleton
 import org.ole.planet.myplanet.datamanager.ApiInterface
 import org.ole.planet.myplanet.datamanager.DatabaseService
-import org.ole.planet.myplanet.model.RealmMyCourse
-import org.ole.planet.myplanet.model.RealmMyLibrary
-import org.ole.planet.myplanet.model.RealmUserModel
+import org.ole.planet.myplanet.data.repository.CourseRepository
+import org.ole.planet.myplanet.data.repository.CourseRepositoryImpl
+import org.ole.planet.myplanet.data.repository.LibraryRepository
+import org.ole.planet.myplanet.data.repository.LibraryRepositoryImpl
+import org.ole.planet.myplanet.data.repository.UserRepository
+import org.ole.planet.myplanet.data.repository.UserRepositoryImpl
 
 @Module
 @InstallIn(SingletonComponent::class)
@@ -43,99 +45,5 @@ object RepositoryModule {
         apiInterface: ApiInterface
     ): CourseRepository {
         return CourseRepositoryImpl(databaseService, apiInterface)
-    }
-}
-
-// User Repository
-interface UserRepository {
-    suspend fun getUserProfile(): String?
-    suspend fun saveUserData(data: String)
-    fun getRealm(): Realm
-    fun getCurrentUser(): RealmUserModel?
-}
-
-class UserRepositoryImpl(
-    private val databaseService: DatabaseService,
-    private val preferences: SharedPreferences,
-    private val apiInterface: ApiInterface
-) : UserRepository {
-
-    override suspend fun getUserProfile(): String? {
-        return preferences.getString("user_profile", null)
-    }
-
-    override suspend fun saveUserData(data: String) {
-        preferences.edit().putString("user_profile", data).apply()
-    }
-
-    override fun getRealm(): Realm {
-        return databaseService.realmInstance
-    }
-
-    override fun getCurrentUser(): RealmUserModel? {
-        return databaseService.realmInstance.where(RealmUserModel::class.java).findFirst()
-    }
-}
-
-// Library Repository
-interface LibraryRepository {
-    fun getAllLibraryItems(): List<RealmMyLibrary>
-    fun getLibraryItemById(id: String): RealmMyLibrary?
-    fun getOfflineLibraryItems(): List<RealmMyLibrary>
-}
-
-class LibraryRepositoryImpl(
-    private val databaseService: DatabaseService,
-    private val apiInterface: ApiInterface
-) : LibraryRepository {
-
-    override fun getAllLibraryItems(): List<RealmMyLibrary> {
-        return databaseService.realmInstance.where(RealmMyLibrary::class.java).findAll()
-    }
-
-    override fun getLibraryItemById(id: String): RealmMyLibrary? {
-        return databaseService.realmInstance.where(RealmMyLibrary::class.java)
-            .equalTo("id", id)
-            .findFirst()
-    }
-
-    override fun getOfflineLibraryItems(): List<RealmMyLibrary> {
-        return databaseService.realmInstance.where(RealmMyLibrary::class.java)
-            .equalTo("resourceOffline", true)
-            .findAll()
-    }
-}
-
-// Course Repository
-interface CourseRepository {
-    fun getAllCourses(): List<RealmMyCourse>
-    fun getCourseById(id: String): RealmMyCourse?
-    fun getEnrolledCourses(): List<RealmMyCourse>
-}
-
-class CourseRepositoryImpl(
-    private val databaseService: DatabaseService,
-    private val apiInterface: ApiInterface
-) : CourseRepository {
-
-    override fun getAllCourses(): List<RealmMyCourse> {
-        return databaseService.realmInstance.where(RealmMyCourse::class.java).findAll()
-    }
-
-    override fun getCourseById(id: String): RealmMyCourse? {
-        return databaseService.realmInstance.where(RealmMyCourse::class.java)
-            .equalTo("courseId", id)
-            .findFirst()
-    }
-
-    override fun getEnrolledCourses(): List<RealmMyCourse> {
-        return databaseService.realmInstance.where(RealmMyCourse::class.java)
-            .equalTo("userId", getCurrentUserId())
-            .findAll()
-    }
-
-    private fun getCurrentUserId(): String {
-        return databaseService.realmInstance.where(RealmUserModel::class.java)
-            .findFirst()?.id ?: ""
     }
 }
