+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/DashboardViewModel.kt
@@ -9,9 +9,9 @@ import java.util.UUID
 import javax.inject.Inject
 import org.ole.planet.myplanet.base.BaseResourceFragment
 import org.ole.planet.myplanet.datamanager.DatabaseService
-import org.ole.planet.myplanet.di.CourseRepository
-import org.ole.planet.myplanet.di.LibraryRepository
-import org.ole.planet.myplanet.di.UserRepository
+import org.ole.planet.myplanet.data.repository.CourseRepository
+import org.ole.planet.myplanet.data.repository.LibraryRepository
+import org.ole.planet.myplanet.data.repository.UserRepository
 import org.ole.planet.myplanet.model.RealmNotification
 import org.ole.planet.myplanet.model.RealmStepExam
 import org.ole.planet.myplanet.model.RealmSubmission
